maintained sku's in database and updated image and archive file year 2 digit to 4 digit and updated production data fetching from shift to latest

+++ src/main/java/TVS_SFL/BearingHousingBackend/dto/BearingHousingProductionDataDTO.java
@@ -19,7 +19,7 @@ public class BearingHousingProductionDataDTO {
     private String barcode;    
     private LocalTime cycleStartTime;    
     private LocalDateTime cycleEndTime;    
-    private Float cycleTime;    
+    private Integer cycleTime;    
     private LocalDateTime productionDateTime;    
     private String shift;    
     private String sku;
